update eth light client with recent changes in authority-binder.ts and light-clients.ts (utils functions)
