+++ sdk/typescript/src/eth-light-client/eth-light-client.ts
@@ -12,7 +12,12 @@ import { bcs } from '../bcs/index.js';
 import { TransactionBlock } from '../builder/index.js';
 import type { DWalletClient } from '../client/index.js';
 import type { Keypair } from '../cryptography/index.js';
-import { getSharedObjectRefById } from '../utils/sui-types.js';
+import {
+	getAuthorityByID,
+	getDWalletBinderByID,
+	getSharedObjectRefById,
+	stringToBcs,
+} from '../utils/light-clients.js';
 import {
 	getBeaconBlockData,
 	getBootstrapData,
@@ -22,12 +27,7 @@ import {
 	getUpdates,
 } from './rpc.js';
 import type { EthereumState } from './utils.js';
-import {
-	getAuthorityBinderByID,
-	getAuthorityByID,
-	getEthereumStateById,
-	stringToArrayU8Bcs,
-} from './utils.js';
+import { getEthereumStateById } from './utils.js';
 
 const packageId = '0x3';
 const ethereumStateModuleName = 'ethereum_authority';
@@ -82,15 +82,15 @@ export const createEthereumAuthority = async (
 
 	let updatesResponseJson = await getUpdates(rpc, syncPeriod);
 	let updatesJson = JSON.stringify(updatesResponseJson.map((update: any) => update['data']));
-	let updatesBcs = stringToArrayU8Bcs(updatesJson);
+	let updatesBcs = stringToBcs(updatesJson);
 
 	let finalityUpdateResponse = await getFinalityUpdate(rpc);
 	let finalityUpdateJson = JSON.stringify(finalityUpdateResponse['data']);
-	let finalityUpdateBcs = stringToArrayU8Bcs(finalityUpdateJson);
+	let finalityUpdateBcs = stringToBcs(finalityUpdateJson);
 
 	let optimisticUpdateResponse = await getOptimisticUpdate(rpc);
 	let optimisticUpdateJson = JSON.stringify(optimisticUpdateResponse['data']);
-	let optimisticUpdateBcs = stringToArrayU8Bcs(optimisticUpdateJson);
+	let optimisticUpdateBcs = stringToBcs(optimisticUpdateJson);
 
 	let stateBcs = bcs.vector(bcs.u8()).serialize(stateBytes, {
 		size: stateBytes.length,
@@ -100,14 +100,12 @@ export const createEthereumAuthority = async (
 
 	// Get Beacon block data for the latest finalized block.
 	let beaconBlockData = await getBeaconBlockData(rpc, finalityUpdateResponse);
-	let beaconBlockTypeBcs = stringToArrayU8Bcs(beaconBlockData.blockType);
-	let beaconBlockBcs = stringToArrayU8Bcs(beaconBlockData.blockJsonString);
-	let beaconBlockBodyBcs = stringToArrayU8Bcs(beaconBlockData.blockBodyJsonString);
-	let beaconBlockExecutionPayloadBcs = stringToArrayU8Bcs(
-		beaconBlockData.blockExecutionPayloadJsonString,
-	);
+	let beaconBlockTypeBcs = stringToBcs(beaconBlockData.blockType);
+	let beaconBlockBcs = stringToBcs(beaconBlockData.blockJsonString);
+	let beaconBlockBodyBcs = stringToBcs(beaconBlockData.blockBodyJsonString);
+	let beaconBlockExecutionPayloadBcs = stringToBcs(beaconBlockData.blockExecutionPayloadJsonString);
 
-	let chainIdentifierBcs = stringToArrayU8Bcs(chainIdentifier);
+	let chainIdentifierBcs = stringToBcs(chainIdentifier);
 
 	const tx = new TransactionBlock();
 	tx.moveCall({
@@ -167,7 +165,7 @@ export const createEthereumSmartContractConfig = async (
 	keypair: Keypair,
 	client: DWalletClient,
 ) => {
-	let networkBcs = stringToArrayU8Bcs(network);
+	let networkBcs = stringToBcs(network);
 
 	const tx = new TransactionBlock();
 	tx.moveCall({
@@ -219,23 +217,21 @@ async function updateEthereumAuthorityState(
 
 	let updatesResponseJson = await getUpdates(consensusRpc, syncPeriod);
 	let updatesJson = JSON.stringify(updatesResponseJson.map((update: any) => update['data']));
-	let updatesBcs = stringToArrayU8Bcs(updatesJson);
+	let updatesBcs = stringToBcs(updatesJson);
 
 	let finalityUpdateResponseJson = await getFinalityUpdate(consensusRpc);
 	let finalityUpdateJson = JSON.stringify(finalityUpdateResponseJson['data']);
-	let finalityUpdateBcs = stringToArrayU8Bcs(finalityUpdateJson);
+	let finalityUpdateBcs = stringToBcs(finalityUpdateJson);
 
 	let optimisticUpdateResponse = await getOptimisticUpdate(consensusRpc);
 	let optimisticUpdateJson = JSON.stringify(optimisticUpdateResponse['data']);
-	let optimisticUpdateBcs = stringToArrayU8Bcs(optimisticUpdateJson);
+	let optimisticUpdateBcs = stringToBcs(optimisticUpdateJson);
 
 	let beaconBlockData = await getBeaconBlockData(consensusRpc, finalityUpdateResponseJson);
-	let beaconBlockTypeBcs = stringToArrayU8Bcs(beaconBlockData.blockType);
-	let beaconBlockBcs = stringToArrayU8Bcs(beaconBlockData.blockJsonString);
-	let beaconBlockBodyBcs = stringToArrayU8Bcs(beaconBlockData.blockBodyJsonString);
-	let beaconBlockExecutionPayloadBcs = stringToArrayU8Bcs(
-		beaconBlockData.blockExecutionPayloadJsonString,
-	);
+	let beaconBlockTypeBcs = stringToBcs(beaconBlockData.blockType);
+	let beaconBlockBcs = stringToBcs(beaconBlockData.blockJsonString);
+	let beaconBlockBodyBcs = stringToBcs(beaconBlockData.blockBodyJsonString);
+	let beaconBlockExecutionPayloadBcs = stringToBcs(beaconBlockData.blockExecutionPayloadJsonString);
 
 	let authoritySharedObjectRef = await getSharedObjectRefById(authorityId, client, true);
 
@@ -310,7 +306,7 @@ export const approveEthereumMessage = async (
 	let currentEthereumStateID = authorityObj?.latest.id as string;
 	let currentEthereumStateObj = await getEthereumStateById(client, currentEthereumStateID);
 
-	let dwalletBinderObj = await getAuthorityBinderByID(dwalletBinderId, client);
+	let dwalletBinderObj = await getDWalletBinderByID(dwalletBinderId, client);
 	let bindToAuthorityObj = dwalletBinderObj?.bind_to_authority;
 
 	let dataSlot = authorityObj?.config.approved_tx_slot as number;
@@ -366,8 +362,8 @@ export const approveEthereumMessage = async (
 
 	let dWalletBinderSharedObjectRef = await getSharedObjectRefById(dwalletBinderId, client);
 	let authoritySharedObjectRef = await getSharedObjectRefById(authorityId, client);
-	let proofBcs = stringToArrayU8Bcs(JSON.stringify(proof));
-	let messageBcs = stringToArrayU8Bcs(message);
+	let proofBcs = stringToBcs(JSON.stringify(proof));
+	let messageBcs = stringToBcs(message);
 
 	const tx = new TransactionBlock();
 	const [messageApprovals] = tx.moveCall({
